requirements.txt
torch==1.8.0
torchvision==0.9.0
pytorch-lightning==1.2.9
opencv-python-headless
numpy
pillow
scikit-image
fastapi
uvicorn[standard]
requests
runpod
easydict
scikit-learn
albumentations==0.5.2
imgaug==0.4.0
omegaconf
hydra-core
lpips
tqdm
torch-fidelity
pandas
scipy
timm

# Added for LaMa runtime compatibility
kornia==0.5.8
webdataset
PyYAML>=6.0


